chore(ts): phase 4d - migrate auth + audio + media hooks
Migrate the heterogeneous auth / audio / media hooks in one pass.
Each is independent and has no deps between files.

- useHfAuth: UseHfAuthResult, typed HF daemon payloads
  (HfAuthStatusPayload, HfOAuthStartPayload). Error narrowing to
  string; `?? null` on optional username/avatar_url.
- useDoA: typed UseDoAResult, debounced isTalking preserved. Coerce
  rawTalking to boolean. Exports getDoADirection / doaToCssRotation
  with strict number|null signatures.
- useAudioAnalyser: typed options + result, AnalyserNode /
  MediaStreamAudioSourceNode refs, Uint8Array<ArrayBuffer> ref to
  satisfy TS 5's stricter getByteFrequencyData signature.
- useWebRTCStream: local minimal ambient typings for window.GstWebRTCAPI
  (producer/session/connection listeners + API ctor options), typed
  StreamStateValue union, UseWebRTCStreamResult. Kept the intentional
  eslint-disable on the auto-connect effect with a comment.

Verified: typecheck, 55 unit tests, lint (no new warnings on touched
files), production build. No behavioural change.

Identified latent bugs (not introduced here, tracked in TODO):
- useHfAuth: mountedRef never reset to true on re-mount -> setState
  ignored after StrictMode / HMR / navigation remount.
- useDoA.getDoADirection: angle = exactly π returns 'left' (π % π = 0).
- useHfAuth: 2s polling has no overlap guard if the daemon is slow.
- useWebRTCStream: session 'error' event typed as Event but JS reads
  e.message (preserved via cast, lib likely emits custom event).

Made-with: Cursor
